NextAuth 설정을 위한 authOptions.ts 파일 추가 및 관련 프로바이더 구성
