+++ src/lib/auth/authOptions.ts
@@ -1,0 +1,58 @@
+// src/lib/auth/authOptions.ts
+import { authApi } from "@/api/auth";
+import type { JoinType } from "@/types/commonUser";
+import { NextAuthOptions } from "next-auth";
+import KakaoProvider from "next-auth/providers/kakao";
+import NaverProvider from "next-auth/providers/naver";
+
+import { createCredentialsProvider } from "./helpers"; // 너가 함수들을 분리한 경우
+
+export const authOptions: NextAuthOptions = {
+  providers: [
+    createCredentialsProvider("user-credentials", "User Credentials", authApi.user.login),
+    createCredentialsProvider("company-credentials", "Company Credentials", authApi.company.login),
+    KakaoProvider({
+      clientId: process.env.KAKAO_CLIENT_ID ?? "",
+      clientSecret: process.env.KAKAO_CLIENT_SECRET ?? "",
+    }),
+    NaverProvider({
+      clientId: process.env.NAVER_CLIENT_ID ?? "",
+      clientSecret: process.env.NAVER_CLIENT_SECRET ?? "",
+    }),
+  ],
+  pages: {
+    signIn: "/auth/login",
+    error: "/auth/login",
+  },
+  session: {
+    strategy: "jwt",
+    maxAge: 30 * 24 * 60 * 60,
+  },
+  callbacks: {
+    async jwt({ token, user }) {
+      if (user) {
+        token.id = user.id;
+        token.sub = user.id;
+        token.name = user.name;
+        token.email = user.email;
+        token.join_type = user.join_type;
+        token.accessToken = user.accessToken;
+        token.refreshToken = user.refreshToken;
+      }
+      return token;
+    },
+    async session({ session, token }) {
+      if (session.user) {
+        session.user.id = token.id as string;
+        session.user.name = token.name as string;
+        session.user.email = token.email as string;
+        session.user.join_type = token.join_type as JoinType;
+        session.accessToken = token.accessToken as string;
+        session.refreshToken = token.refreshToken as string;
+      }
+      return session;
+    },
+  },
+  secret: process.env.NEXTAUTH_SECRET,
+  debug: process.env.NODE_ENV === "development",
+};
